Enable writable http transports.
* bzrlib/transport/http/__init__.py:
(HttpTransportBase.abspath):  Forbids  directories  only when  the
transport is read-only (I think it was the original intent), as it
blocks all attempts to write a plugin that allows writing on http.
(HttpServer.__init__):  Allows  subclasses  to specify  their  own
request handler.

* bzrlib/tests/test_transport_implementations.py:
(TransportTests.test_put): Cast to StringIO.
(TransportTests.test_rmdir): Use has to  test the existence of the
file  instead  of  stat.  A  transport  may  be  writable  without
implementing stat.  The intent  of the test  is to test  rmdir not
stat.