Fix the 'init connects multiple times' in a different way.
* transport/__init__.py:
(get_transport): Make transports default to None instead of [],
it's easier and less confusing in its other uses.

* tests/test_bzrdir.py:
(TestBzrDirFormat): Remove the test, the new fix is harder to test
for. What we really want to test is that *connections* are reused,
LocalTransport.clone create new objects anyway, so there is no way
to verify that a LocalTransport is really reused. So we must
instrument transports that use connection to be able to catch
multiple connections.

* bzrdir.py:
(BzrDir.create): Add a transports parameter for reuse.
(BzrDir.create_branch_convenience): Add a transports parameter for
reuse.
(BzrDirFormat.initialize): Add a transports parameter for reuse.

* builtins.py:
(cmd_init.run): Pass the known transport to
create_branch_convenience in its own list.