Make version-info --custom imply --all.
When using a custom template you can use variables like {clean} which
are not enabled by default. This causes an ugly traceback.

This fixes the issue by implying --all when --custom (or more exactly
--template) is used, so that all variables are available in templates.

This was the easier fix than having the template code report that the
user should pass --check-clean to have {clean} available, but will
have a performance impact on all users of version-info --custom,
regardless of whether they use {clean}.