Search from all of the known revisions.
This lets us build up the numbering for all the stuff we know, and
allows us to stop early if there is convergence from a known revision
to the one in question.
In 'easy' cases this will search more than it has to, but it prevents
the worst-case behavior of always searching to the NULL revision.