Add a couple of special cases to urlutils._win32_path_(from|to)_url
 * 'file://' resolves to a path of ''
 * '/' resolves to a URI of 'file://'

This fixes #240910 by allowing ``bzr serve --directory=/`` to access all paths
on the Windows filesystem rather than just paths relative to the root of the 
drive Python.exe is running from. This is the value for the directory arg
used by remote SSH invocation, so the most desirable effect is to allow
you to reach any permitted path on the filesystem using SSH as with *nix.

Internally, this works by creating a LocalTransport instance with an empty string 
as a base path. This allows you to pass normal (for Windows) ssh/sftp URIs in the style :

bzr://camelot/x:/foo/bar
bzr+ssh://robin@camelot/x:/foo/bar