Make TreeTransform update the inventory with new kind information.
If an entry simply changes kind in a TreeTransform it wouldn't previously
be updated in the inventory, leading to a dirstate that didn't match what
was on disk.

This change makes TreeTransform keep track of kind changes and add an
inventory_delta entry to change the kind of the entry.

It is only implemented for ->directory changes currently.