Update to do a single iteration over the chunks.
This costs 600us versus 430us for the case where the object is
already a list of lines. However it is only 1.2ms rather than 3ms
when everything is in a single buffer.

The biggest advantage is that 'chunks' *could* be an iterator,
rather than requiring it to already have all the results.