Change the serialized form for leaf nodes.
Instead of storing the raw bytes, we store a 'number of lines' counter,
and then that many lines of value.
This allows us to have key:value pairs, where the value contains newlines.
This breaks a lot of tests, because it changes the serialized form.
At this point, though, the serialization matches the deserialization.