Small optimization for iter_changes().
The 'interesting' work is still being done via 'heapq.push/pull'.
However, the very common case is that the root node on both sides
is an InternalNode with practically identical structure.
So instead of doing a heappush and then having to walk both sides
in sync, and see if everything lines up, etc. Do the culling ahead
of time.
This takes us from 11s+ down to 9.3s.
Interestingly, it only hits 1k times out of 13k calls to iter_changes,
but still has that significant of an improvement.
This means that the bug with item prefix probably is causing us to
page in far too many nodes.