Further understanding shows just how unlikely a collision is, at 8 bytes and 100M keys.
So next I need to revise the recommended size of various objects.
The mini-index size should be tuned to be reasonable versus the
expected number of bytes that need to be read to resolve the
reference. And shrinking the keys changes the width of the bytes,
and effects the chance of collision.
Very likely we can write an index that is <10 bytes per entry.