Move even more functionality into EquivalenceTable.
At this point, ET takes care of the delta generation, which is nice.
Also, I uncovered a bug in how we handle insert of > 1 line, because
the matcher is strictly line based, but the insert was based on
number of bytes. So it created a 'line' which could never be
matched.
We just need to tie in 'make_delta' and we can run most of the
tests, and then start working on apply_delta()