Split out the auto-upgrade logic inside CreateRepository.acquire_repository.
Instead of always upgrading the repo format *and* the branch format, now we only upgrade
the repo if it doesn't support stacking, and only upgrade the branch if *it*
doesn't support stacking.
It is a bit of an end-run around the fact that initialize_on_transport_ex
tries to set a repository format but *doesn't* set a branch format, and then
goes on to create a repo which wants to know what branch format it would create.
Which causes it to revert to the default branch format which doesn't support
stacking.

At least it works.