Add the time_graph code for benchmarking.
Add a lot of debugging counters, etc.

At the moment, 'random sampling' shows a performance improvement of 4:1
however, just checking 'parents of a merge' in bzr.dev shows a
performance *loss* of about 1:2. So still need to do a bit more
probing to figure out why. It may be the overhead of using objects,
or it may be an issue of the algorithm.