Change the KnownGraph.merge_sort api.
Instead of being a tuple api, it is now an Object api. TIMEIT testing at least
seems to say that _MergeSortNode.key is no slower than using a tuple.
Allocating the nodes is also not much slower (especially since we already needed
them.)

In the end it makes for an api which, IMO, is much easier to use, so
since performance isn't impacted, it is worth switching.


Also, determined that we needed to implement the new get_known_graph... on
the weave code.