Remove suppression of alternate switches for the same option.
Previous versions did try to remove multiple switches coming from the same
option.  The feature was already present in the static completion script. 
The options to be removed, however, were not based on the command about to
get completed.  WIth the full list of all switches provided by all options,
the duplicate replacement code became rather long.

For the moment, this patch drops the whole removal stuff. Reasons for this:
* There is no guarantee that multiple switches from the same option are
  always undesirable.
* There duplicate removal code got quite big for little gain.
* The corresponding python code made reading the function harder than
  necessary otheriwse.

If there is ever any kind of duplicate removal in the future, it would get
implemented in a way that is specific to the bzr command in question, and
would take the class of the option into account to ensure that duplicates
are indeed pointless.