Some comment cleanup, implement a special case for Py_EQ when both objects are intrened.
btree_mem_test.py doesn't show a significant impact. Real-world we probably
only use Py_EQ comparisons in sets and dicts when we get conflicts, also
note that we *don't* do the comparison for StaticTuple_Intern() because the
objects haven't been interned yet...

However, it seems 'correct' to do the check, so I'll leave it in.