add a hack-around to handle differences between pyrex 0.9.6 and 0.9.8
Namely, 0.9.8 creates a function 'import__simple_set_pyx' and tries to import
the module as '_simple_set_pyx', while 0.9.8 creates the function
'import_bzrlib___simple_set_pyx', and imports it as 'bzrlib._simple_set_pyx'.
The latter seems to be the preferred way forward, so we hack up sys.modules
so that we can get at it there.
It is a bit ugly, but I've tested that it works under both versions.