Fix bug #44062, Always suppress deprecation warnings.
Python 2.6 now 'comes with' a deprecation warning suppression as part of
the threading module. So the current check that says 'if someone manually
configured a deprecation warning filter, don't add ours' now breaks
because there is always one present.
So instead, we just always suppress all deprecation warnings if we are in
a 'final' release.