Added GPL 2 license document and copyright notice.
The wording of the copyright notice in the source files is actually based on
the one suggested by GPL 3, but with the version number modified to 2.  This
avoids reference to the postal address in favour of a web URL.  The license
wasn't completely migrated to GPL 3 in order to allowing inclusion into a
GPL2-only code tree like bzr itself.