Add --plugin switch to selectively include plugins.
This is intended for scenarios where a bash completion script is generated
once but distributed to a number of sifferent bzr setups.  In this case the
list of plugins for which commands are included in the completion function
can be controlled and used to restrict the function to a fixed set of
plugins.  Other aspects of command line completion, in particular registry
values registered by plugins, are not affected by this switch.