The message, 'Using saved push location: ...' reported by 'bzr push' without a location (but a 'remembered' push location in .bzr/branch/branch.conf) is now supressed when the -q flag is used.  This is consistent with the behaviour of bzr pull.
Fix to bug lp:803053 as described below:

Issuing the command 'bzr push -q' without a location (and a 'remembered' push location in .bzr/branch/branch.conf) reports the message 'Using saved push location: ...', unlike 'bzr pull -q' which is quiet. I would expect this message to be suppressed when the -q flag is used, as it is neither an error or warning.