More refactoring.
Rewrite the tests to focus on extracting the content vs reporting the content,
and then rewrite to focus on the permutations that matter.
I think we actually have more tests now, because display tests still
need to test all the permutations, and now we add a couple more for
getting the data out of the branch.
Though the display tests do run faster...