Instrumentation is revealing.
Under lsprof it takes 21s to do the fetch with depth=100 and size=100. We have 16 calls to
limited_search_result_from_parent, and only 1 call to 'search_result_from_parent_map'.
What is especially revealing is that 13.7s is spent in the RPC, which means that
depth=100 is expensive on the *server* side, not the client side.

Also interesting is:
5.6802  bzrlib.graph:1986(limited_search_result_from_parent_map)
1.9023  +bzrlib.graph:1937(_find_possible_heads)
2.6919  +bzrlib.graph:2017(first_search)
0.0928  +bzrlib.graph:1317(get_result)
0.9129  +bzrlib.graph:2021(second_search)

For some reason, the search with fewer heads is a lot faster than the original wide search.
Probably an inefficiency with BreadthFirstSearcher.
_find_possible_heads can also probably be tuned:
1.9023  bzrlib.graph:1937(_find_possible_heads)
0.5954  +<method 'difference_update' of 'set' objects>
0.4819  +bzrlib.graph:1921(invert_parent_map)
0.3779  +<method 'update' of 'set' objects>
0.0006  +<method 'add' of 'set' objects>

invert_parent_map doesn't dominate, and we have a lot of calls to difference_update, etc.