Rename --include-merges option to --include-sidelines.
The rationale behind this rename is the following: a MERGE is a single
commit with two parents.  But instead of the merge itself being ommitted
from the log, it is the line(s) of development starting with the secondary
parent(s).  We already call the primary undotted-revno left-parent-only line
of development MAINLINE, so calling the other lines SIDELINES should be an
intuitive concept.

The UI change is backwards compatible, as the old option name is preserved. 
If both options are specified, in either their positive or their negative
form, then the sideline one takes precedence.  Conflicting options are not
reported.  That means that a --include-sidelines argument in an alias might
take precedence over an --no-include-merges argument on the command line. 
As aliases and command line tend to be written by the same person, this
should be a rare scenario.

The tests have not been updated yet, in order to show that the modification
does not break any existing tests.  Updated tests will be in a separate
commit.

The Python API hasn't been changed, as it's more difficult to hide
backwards-compatibility options there, and as two options to the same effect
might confuse programmers.