Now we've implemented the basic structure.
I tested it manually, and it works as we wanted, the server waits for the current
request to stop, but keeps running until it does so. Once the current request
has finished, it closes the connection to the client.

There are now a couple of race conditions in the test suite. I caught a couple
of them, like the connection thread was being started before we added it to
the list of active connections.
The next race that I know of, is the test suite expects that a request&response
finish is enough for us to have run _poll_active_connections.
However, there is no actual interlocking there.
So we actually need to do a new connection, which may give us 1-or-2 active
connections, but then we can check if the old handler is still around.