Added test case to reproduce error (test_mv_dirs_non_ascii).
Test added to blackbox/test_commit.py.  The bug is in fact in dirstate,
but I couldn't successfully add a test to test_dirstate.py; there's
something wrong with expected encodings in TestUpdateBasisByDelta, which
is where this test belongs.

The blackbox test is based on Robert Collins' foo.sh [1], which is:

    rm -rf test
    bzr init test
    mkdir test/abcÃ„Ã–Ãœabc
    touch test/abcÃ„Ã–Ãœabc/bar
    bzr add test 
    bzr ci -m "import to svn" test
    bzr mv test/abcÃ„Ã–Ãœabc test/abc
    bzr commit -m "fails" test

[1] https://bugs.launchpad.net/bzr/+bug/185211/+attachment/195003/+files/foo.sh