Fix a hang in the bundle tests.
The dummy server we implemented only allowed a single connection, and then it
would leave the main socket in 'listen' without ever calling accept again.
If we close it early, then we get ConnectionError instead of ConnectionReset
and we were trynig to test Reset.
So we just loop on the connect.
In bzr-2.5 it is done differently because we have the very nice test_server
infrastructure.